PET-79 Separated UserService from DatabaseService
