+++ src/main/java/org/controllers/ShelterManagerController.java
@@ -17,6 +17,7 @@ import javafx.stage.Stage;
 import org.model.Pet;
 import org.model.User;
 import org.services.DatabaseService;
+import org.services.UserService;
 
 import javax.xml.crypto.Data;
 import java.io.File;
@@ -130,7 +131,7 @@ public class ShelterManagerController {
         crt.setInfo(petInfo.getText());
         crt.setImagePath(imagePath);
         user.addPet(crt);
-        DatabaseService.updateUser(user);
+        UserService.updateUser(user);
 
         Node node = (Node) event.getSource();
         Stage currentStage = (Stage) node.getScene().getWindow();
@@ -159,7 +160,7 @@ public class ShelterManagerController {
             }
         }
 
-        DatabaseService.updateUser(user);
+        UserService.updateUser(user);
 
         this.updateList();
     }
